requirements.txt
Django
djangorestframework
Pillow
opencv-python
numpy

pip freeze > requirements.txt